Fixing AMD64 builds

+++ include/machine/_types.h
@@ -54,8 +54,8 @@ typedef	short			__int16_t;
 typedef	unsigned short		__uint16_t;
 typedef	int			__int32_t;
 typedef	unsigned int		__uint32_t;
-typedef	long			__int64_t;
-typedef	unsigned long		__uint64_t;
+typedef	long long			__int64_t;
+typedef	unsigned long long		__uint64_t;
 
 /*
  * Standard type definitions.
